Improvements to the secure-session tutorial based on Mark Spielman's second review.
