+++ docs/_config.yml
@@ -13,7 +13,7 @@ docs-core-concepts: http://docs.solace.com/Features/Core-Concepts.htm
 links-downloads: http://dev.solace.com/downloads/
 links-pubsub-tutorial: http://dev.solace.com/get-started/java-tutorials/publish-subscribe_java/
 links-tls-client: http://docs.solace.com/Solace-Messaging-APIs/Creating-Secure-Sessions.htm
-links-tls-server: http://docs.solace.com/Configuring-and-Managing-Routers/Managing-TLS-SSL-Service.htm
+links-tls-server: https://docs.pivotal.io/solace-messaging/installing.html#configure
 links-vcap: http://docs.pivotal.io/solace-messaging/credentials.html
 #----------------------------------------------------
 # Generally don't modify items below this point
