+++ docs/_data/tutorials.yml
@@ -2,4 +2,4 @@
   tutorials:
   - java-app
   - spring-cloud
-  - secure-sessions
+  - secure-session
