+++ secure-session/build.gradle
@@ -2,7 +2,7 @@ buildscript {
     ext {
         springBootVersion = '1.3.3.RELEASE'
         springCloudVersion = '1.2.3.RELEASE'
-        solaceJCSMPVersion = '+'
+        solaceJCSMPVersion = '10.+'
         solaceCloudVersion = '0.1+'		
     }
     repositories {
@@ -19,15 +19,12 @@ apply plugin: 'idea'
 apply plugin: 'spring-boot'
 
 jar {
-    baseName = 'solace-sample-spring-cloud'
+    baseName = 'solace-sample-secure-session'
     version =  ''
 }
 
 repositories {
     mavenCentral()
-    flatDir {
-       dirs '../libs'
-    }
 }
 
 dependencies {
@@ -45,9 +42,6 @@ dependencies {
     compile("com.solace.labs.cloud.cloudfoundry:solace-labs-spring-cloud-connector:${solaceCloudVersion}")
     
     // Solace Java API
-    compile("commons-lang:commons-lang:2.6")
-    compile("commons-logging:commons-logging:1.2")
-    compile("com.solacesystems:sol-common:${solaceJCSMPVersion}")
     compile("com.solacesystems:sol-jcsmp:${solaceJCSMPVersion}")
 }
