+++ secure-session/manifest.yml
@@ -2,6 +2,6 @@
 applications:
 - name: solace-sample-secure
   memory: 512M
-  path: build/libs/solace-sample-spring-cloud.jar
+  path: build/libs/solace-sample-secure-session.jar
   services:
   - solace-messaging-sample-instance
