+++ secure-session/src/main/java/com/solace/samples/cloudfoundry/securesession/model/SimpleMessage.java
@@ -16,32 +16,31 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-package com.solace.samples.cloudfoundry.springcloud.model;
+package com.solace.samples.cloudfoundry.securesession.model;
 
 public class SimpleMessage {
-	
-	private String topic;
-	private String body;
-	
-	public SimpleMessage() {
-		this.topic = "";
-		this.body = "";
-	}
-	
-	public String getTopic() {
-		return topic;
-	}
-	
-	public void setTopic(String topic) {
-		this.topic = topic;
-	}
 
-	public String getBody() {
-		return body;
-	}
-	
-	public void setBody(String body) {
-		this.body = body;
-	}
+    private String topic;
+    private String body;
+
+    public SimpleMessage() {
+        this.topic = "";
+        this.body = "";
+    }
+
+    public String getTopic() {
+        return topic;
+    }
+
+    public void setTopic(String topic) {
+        this.topic = topic;
+    }
+
+    public String getBody() {
+        return body;
+    }
+
+    public void setBody(String body) {
+        this.body = body;
+    }
 }
