+++ settings.gradle
@@ -1,1 +1,1 @@
-include "java-app", "secure-app", "spring-cloud"
+include "java-app", "secure-session", "spring-cloud"
